accessible/tests/mochitest/tree/test_svg.html

<!DOCTYPE html>
<html>
<head>
  <title>SVG Tree Tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest() {
      var accTree = {
        role: ROLE_DIAGRAM,
        children: [
          {
            role: ROLE_TEXT_CONTAINER,
            children: [
              {
                role: ROLE_TEXT_LEAF,
              },
            ],
          },
        ],
      };
      testAccessibleTree("svgItem", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
  <svg id="svgItem">
    <text>This is some text</text>
  </svg>
</body>
</html>
